Refactor model import/export API
Closes #1505
