+++ workcraft/CircuitPlugin/src/org/workcraft/plugins/circuit/tasks/SpotChainTask.java
@@ -127,7 +127,7 @@ public class SpotChainTask implements Task<SpotChainOutput> {
             Stg sysModStg = StgUtils.loadOrImportStg(sysStgFile);
             File sysModStgFile = new File(directory, StgUtils.SYSTEM_FILE_PREFIX + StgUtils.MODIFIED_FILE_SUFFIX + stgFileExtension);
             FileOutputStream sysModStgStream = new FileOutputStream(sysModStgFile);
-            SerialiserUtils.writeModel(sysModStg, sysModStgStream, SerialiserUtils.Style.STG, true);
+            SerialiserUtils.writeModel(sysModStg, sysModStgStream, sysModStgFile, SerialiserUtils.Style.STG, true);
             sysModStgStream.close();
             monitor.progressUpdate(0.30);
