+++ workcraft/FstPlugin/src/org/workcraft/plugins/fst/interop/SgImporter.java
@@ -20,21 +20,22 @@ public class SgImporter implements Importer {
     }
 
     @Override
-    public ModelEntry importFrom(InputStream in, String serialisedUserData) throws DeserialisationException {
-        return new ModelEntry(new FstDescriptor(), importSG(in));
+    public ModelEntry deserialise(InputStream in, String serialisedUserData) throws DeserialisationException {
+        return new ModelEntry(new FstDescriptor(), deserialiseFst(in));
     }
 
-    public Fst importSG(InputStream in) throws DeserialisationException {
+    public static Fst deserialiseFst(InputStream in) throws DeserialisationException {
+        SgParser parser = new SgParser(in);
+        if (DebugCommonSettings.getParserTracing()) {
+            parser.enable_tracing();
+        } else {
+            parser.disable_tracing();
+        }
         try {
-            SgParser parser = new SgParser(in);
-            if (DebugCommonSettings.getParserTracing()) {
-                parser.enable_tracing();
-            } else {
-                parser.disable_tracing();
-            }
             return parser.parse();
         } catch (FormatException | ParseException e) {
             throw new DeserialisationException(e);
         }
     }
+
 }
