+++ workcraft/PetrifyPlugin/src/org/workcraft/plugins/petrify/tasks/WriteSgConversionTask.java
@@ -4,7 +4,7 @@ import org.workcraft.Framework;
 import org.workcraft.exceptions.NoExporterException;
 import org.workcraft.interop.Exporter;
 import org.workcraft.plugins.fst.Fst;
-import org.workcraft.plugins.fst.utils.FstUtils;
+import org.workcraft.plugins.fst.interop.SgImporter;
 import org.workcraft.plugins.petri.PetriModel;
 import org.workcraft.plugins.stg.interop.StgFormat;
 import org.workcraft.tasks.ProgressMonitor;
@@ -16,6 +16,7 @@ import org.workcraft.utils.WorkspaceUtils;
 import org.workcraft.workspace.WorkspaceEntry;
 
 import javax.swing.*;
+import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
@@ -96,7 +97,7 @@ public class WriteSgConversionTask implements Task<WriteSgConversionOutput> {
 
                 ExternalProcessOutput output = result.getPayload();
                 if (result.isSuccess()) {
-                    Fst fst = FstUtils.importFst(output.getStdout());
+                    Fst fst = SgImporter.deserialiseFst(new ByteArrayInputStream(output.getStdout()));
                     return Result.success(new WriteSgConversionOutput(output, fst));
                 }
