+++ workcraft/StgPlugin/src/org/workcraft/plugins/stg/interop/LpnExporter.java
@@ -1,11 +1,13 @@
 package org.workcraft.plugins.stg.interop;
 
-import org.workcraft.interop.AbstractSerialiseExporter;
-import org.workcraft.plugins.stg.serialisation.LpnSerialiser;
+import org.workcraft.dom.Model;
+import org.workcraft.interop.Exporter;
+import org.workcraft.plugins.petri.PetriModel;
+import org.workcraft.plugins.stg.serialisation.SerialiserUtils;
 
-public class LpnExporter extends AbstractSerialiseExporter {
+import java.io.OutputStream;
 
-    private final LpnSerialiser serialiser = new LpnSerialiser();
+public class LpnExporter implements Exporter {
 
     @Override
     public LpnFormat getFormat() {
@@ -13,8 +15,13 @@ public class LpnExporter extends AbstractSerialiseExporter {
     }
 
     @Override
-    public LpnSerialiser getSerialiser() {
-        return serialiser;
+    public boolean isCompatible(Model model) {
+        return model instanceof PetriModel;
+    }
+
+    @Override
+    public void serialise(Model model, OutputStream out) {
+        SerialiserUtils.writeModel(model, out, getCurrentFile(), SerialiserUtils.Style.LPN, true);
     }
 
 }
