+++ workcraft/WorkcraftCore/src/org/workcraft/plugins/builtin/interop/SvgExporter.java
@@ -11,14 +11,6 @@ import java.io.OutputStream;
 public class SvgExporter implements Exporter {
 
     @Override
-    public void exportTo(Model model, OutputStream out) throws SerialisationException {
-        if (!(model instanceof VisualModel)) {
-            throw new SerialisationException("Non-visual model cannot be exported as SVG file.");
-        }
-        BatikUtils.transcode((VisualModel) model, out);
-    }
-
-    @Override
     public boolean isCompatible(Model model) {
         return model instanceof VisualModel;
     }
@@ -26,6 +18,11 @@ public class SvgExporter implements Exporter {
     @Override
     public SvgFormat getFormat() {
         return SvgFormat.getInstance();
+    }
+
+    @Override
+    public void serialise(Model model, OutputStream out) throws SerialisationException {
+        BatikUtils.transcode((VisualModel) model, out);
     }
 
 }
